pyproject.toml
[project]
name = "onboard-gnn"
version = "0.1.0"
description = "Simplified ML training pipeline for new engineers - GNN baselines on DeepJEB"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.0",
    "torch-geometric>=2.4",
    "lightning>=2.0",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "numpy>=1.24",
    "pandas>=2.0",
    "click>=8.0",
    "rich>=13.0",
    "torchmetrics>=1.0",
    "wandb>=0.15",
    "mlflow-skinny>=3.8.0",
    "python-dotenv>=1.0",
    "h5py>=3.0",
    "pyarrow>=22.0.0",
    "matplotlib>=3.10.8",
]

[dependency-groups]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["shared", "models", "tasks"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["shared", "models", "tasks"]
python_files = ["*_test.py"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
